// frontend/src/api/generated/model/highlightLabel.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * crossbill API
 * OpenAPI spec version: 0.0.0
 */

/**
 * Resolved label info for a highlight.
 */
export interface HighlightLabel {
  /** ID of the highlight style */
  highlight_style_id?: number | null;
  /** Resolved label text */
  text?: string | null;
  /** Resolved UI color */
  ui_color?: string | null;
}